this adds option to configure S3 style access (path-style URLs or virtual hosted-style URLs) with the environment variable AWS_ENABLE_PATH_STYLE_ACCESS.
Note that when this variable isn't provided the default hard-coded value is used depending on the module. With this environment variable one is able to override the default value. https://github.com/instaclustr/esop/issues/53 provides some more context.

Besides this option the Ceph module is updated to use the builder in configuring/providing AmazonS3.